move winterfell field to fields dir
