+++ math/src/field/fields/mod.rs
@@ -9,3 +9,7 @@ pub mod p448_goldilocks_prime_field;
 pub mod u64_goldilocks_field;
 /// Implementation of prime fields over 64 bit unsigned integers.
 pub mod u64_prime_field;
+
+/// Winterfell and miden field compatibility
+#[cfg(feature = "winter_compatibility")]
+pub mod winterfell;
