+++ math/src/field/mod.rs
@@ -10,6 +10,3 @@ pub mod test_fields;
 /// Common behaviour for field elements.
 pub mod traits;
 
-/// Winterfell and miden field compatibility
-#[cfg(feature = "winter_compatibility")]
-pub mod winterfell;
