+++ math/src/traits.rs
@@ -1,9 +1,3 @@
-#[cfg(feature = "winter_compatibility")]
-use miden_core::Felt;
-
-#[cfg(feature = "winter_compatibility")]
-use winter_math::FieldElement as WinterFieldElement;
-
 use crate::{
     errors::DeserializationError,
     field::{element::FieldElement, traits::IsField},
@@ -31,39 +25,6 @@ pub trait ByteConversion {
     fn from_bytes_le(bytes: &[u8]) -> Result<Self, ByteConversionError>
     where
         Self: Sized;
-}
-
-#[cfg(feature = "winter_compatibility")]
-impl ByteConversion for Felt {
-    fn to_bytes_be(&self) -> Vec<u8> {
-        Felt::elements_as_bytes(&[self.clone()]).to_vec()
-    }
-
-    fn to_bytes_le(&self) -> Vec<u8> {
-        Felt::elements_as_bytes(&[self.clone()]).to_vec()
-    }
-
-    fn from_bytes_be(bytes: &[u8]) -> Result<Self, ByteConversionError>
-    where
-        Self: Sized,
-    {
-        unsafe {
-            let res = Felt::bytes_as_elements(bytes)
-                .map_err(|_| ByteConversionError::FromBEBytesError)?;
-            Ok(res[0])
-        }
-    }
-
-    fn from_bytes_le(bytes: &[u8]) -> Result<Self, ByteConversionError>
-    where
-        Self: Sized,
-    {
-        unsafe {
-            let res = Felt::bytes_as_elements(bytes)
-                .map_err(|_| ByteConversionError::FromBEBytesError)?;
-            Ok(res[0])
-        }
-    }
 }
 
 /// Serialize function without args
